// frontend/src/app/api/models/company-characteristic-read-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CompanyCharacteristicReadDto {
  characteristicId?: string | null;
  color?: string | null;
  icon?: string | null;
  id?: string;
  label?: string | null;
  unit?: string | null;
  value?: number;
}